refactor(core): migrate properties subsystem to move semantics and static reflection
- Replace `std::shared_ptr` with `std::unique_ptr` in internal `Properties` storage to enforce clean ownership and zero-overhead.
- Remove global `Type` enum and the monolithic conversion switch in `PropertyValue` to eliminate tight coupling.
- Delegate text formatting and type metadata (`typeName`) to the extensible `PropertySerializer<T>` trait, satisfying the Open/Closed Principle.
- Implement the polymorphic clone pattern (`PropertyBase::clone`) to natively support safe deep copies and efficient moves in `Properties`.
- Update `setProperty` using C++20 `std::decay_t` for cleaner and optimized type-safe instantiation.

+++ test/featmatch/kaze/kaze_test.cpp
@@ -202,7 +202,7 @@ BOOST_FIXTURE_TEST_CASE(kaze_detector_properties, KazeDetectorDescriptorTest)
 
     BOOST_CHECK_EQUAL(true, properties.getProperty<bool>("ExtendedDescriptor"));
     BOOST_CHECK_EQUAL(true, properties.getProperty<bool>("UprightDescriptor"));
-    BOOST_CHECK_EQUAL(0.05f, properties.getProperty<double>("Threshold"));
+    BOOST_CHECK_EQUAL(0.05f, properties.getProperty<float>("Threshold"));
     BOOST_CHECK_EQUAL(8, properties.getProperty<int>("Octaves"));
     BOOST_CHECK_EQUAL(3, properties.getProperty<int>("OctaveLayers"));
     BOOST_CHECK_EQUAL("DIFF_PM_G1", properties.getProperty<std::string>("Diffusivity"));
